--- projects/npm-tools/packages/npm-scripts/src/presets/standard/dependencies/clay.js
@@ -1,37 +1,0 @@
-/**
- * SPDX-FileCopyrightText: © 2019 Liferay, Inc. <https://liferay.com>
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-module.exports = [
-	'clay-alert',
-	'clay-autocomplete',
-	'clay-badge',
-	'clay-button',
-	'clay-card',
-	'clay-card-grid',
-	'clay-checkbox',
-	'clay-collapse',
-	'clay-component',
-	'clay-data-provider',
-	'clay-dataset-display',
-	'clay-dropdown',
-	'clay-icon',
-	'clay-label',
-	'clay-link',
-	'clay-list',
-	'clay-loading-indicator',
-	'clay-management-toolbar',
-	'clay-modal',
-	'clay-multi-select',
-	'clay-navigation-bar',
-	'clay-pagination',
-	'clay-pagination-bar',
-	'clay-portal',
-	'clay-progress-bar',
-	'clay-radio',
-	'clay-select',
-	'clay-sticker',
-	'clay-table',
-	'clay-tooltip',
-];
