--- projects/npm-tools/packages/npm-scripts/src/presets/standard/dependencies/liferay.js
@@ -1,13 +1,0 @@
-/**
- * SPDX-FileCopyrightText: © 2019 Liferay, Inc. <https://liferay.com>
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-module.exports = [
-	'frontend-js-metal-web',
-	'frontend-js-react-web',
-	'frontend-js-web',
-	'frontend-taglib-clay',
-	'frontend-taglib',
-	'hello-soy-web',
-];
