--- projects/npm-tools/packages/npm-scripts/src/presets/standard/dependencies/metal.js
@@ -1,40 +1,0 @@
-/**
- * SPDX-FileCopyrightText: © 2019 Liferay, Inc. <https://liferay.com>
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-module.exports = [
-	'metal',
-	'metal-affix',
-	'metal-ajax',
-	'metal-anim',
-	'metal-aop',
-	'metal-assertions',
-	'metal-clipboard',
-	'metal-component',
-	'metal-debounce',
-	'metal-dom',
-	'metal-drag-drop',
-	'metal-events',
-	'metal-incremental-dom',
-	'metal-jsx',
-	'metal-key',
-	'metal-keyboard-focus',
-	'metal-multimap',
-	'metal-pagination',
-	'metal-path-parser',
-	'metal-position',
-	'metal-promise',
-	'metal-router',
-	'metal-scrollspy',
-	'metal-soy',
-	'metal-soy-bundle',
-	'metal-state',
-	'metal-storage',
-	'metal-structs',
-	'metal-throttle',
-	'metal-toggler',
-	'metal-uri',
-	'metal-useragent',
-	'metal-web-component',
-];
